fixed 3 or more
working now
there are cases where it will not find a pathway between 3 or more.
this is because it has to go back on its self.
this could be fixed by calling init again. the positions would have to be saved though or else it would be overwritten. this might decrease performance