fixed heap error, (i think)
when calling deconstructor, poiints 1d dynamic array was empty anyway so would crash on that when deconstructor was called in "regenerate".

also forgot to delete the incr arrray in vd function.

also added a break in phase 1 if the int it goes too high which for some reason produces an error